[OF-1057] refac: split component tests out
- Split component related tests into their own files
