+++ Tests/src/PublicAPITests/MultiPlayerTests.cpp
@@ -17,17 +17,7 @@
 #include "Awaitable.h"
 #include "CSP/CSPFoundation.h"
 #include "CSP/Common/Optional.h"
-#include "CSP/Multiplayer/Components/AnimatedModelSpaceComponent.h"
-#include "CSP/Multiplayer/Components/AudioSpaceComponent.h"
-#include "CSP/Multiplayer/Components/CollisionSpaceComponent.h"
-#include "CSP/Multiplayer/Components/CustomSpaceComponent.h"
-#include "CSP/Multiplayer/Components/ExternalLinkSpaceComponent.h"
-#include "CSP/Multiplayer/Components/FogSpaceComponent.h"
-#include "CSP/Multiplayer/Components/ImageSpaceComponent.h"
-#include "CSP/Multiplayer/Components/ReflectionSpaceComponent.h"
-#include "CSP/Multiplayer/Components/ScriptSpaceComponent.h"
 #include "CSP/Multiplayer/Components/StaticModelSpaceComponent.h"
-#include "CSP/Multiplayer/Components/VideoPlayerSpaceComponent.h"
 #include "CSP/Multiplayer/MultiPlayerConnection.h"
 #include "CSP/Multiplayer/ReplicatedValue.h"
 #include "CSP/Multiplayer/SpaceEntity.h"
@@ -46,10 +36,8 @@
 #include "signalrclient/signalr_value.h"
 
 #include "gtest/gtest.h"
-#include <CSP/Multiplayer/Components/ConversationSpaceComponent.h>
+#include <CSP/Multiplayer/Components/ImageSpaceComponent.h>
 #include <CSP/Multiplayer/Components/LightSpaceComponent.h>
-#include <CSP/Multiplayer/Components/PortalSpaceComponent.h>
-#include <CSP/Multiplayer/Components/SplineSpaceComponent.h>
 #include <atomic_queue/defs.h>
 #include <chrono>
 #include <filesystem>
@@ -1332,6 +1320,7 @@ CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, NetworkEventMultiTypeTest)
 
 #if RUN_MULTIPLAYER_INTERACTIVEMOVEMENT_TEST
 	#include <conio.h>
+	#include <CSP/Multiplayer/Components/LightSpaceComponent.h>
 
 csp::multiplayer::MultiplayerConnection* Connection;
 
@@ -1628,193 +1617,7 @@ CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, ConnectionInterruptTest)
 }
 #endif
 
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_USE_PORTAL_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, UsePortalTest)
-{
-	SetRandSeed();
 
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	const char* TestSpaceName2		  = "OLY-UNITTEST-SPACE-REWIND-2";
-	const char* TestSpaceDescription2 = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	char UniqueSpaceName2[256];
-	SPRINTF(UniqueSpaceName2, "%s-%s", TestSpaceName2, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	csp::systems::Space Space2;
-	CreateSpace(SpaceSystem, UniqueSpaceName2, TestSpaceDescription2, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space2);
-
-	csp::common::String PortalSpaceID;
-
-	const csp::common::String UserName		= "Player 1";
-	const SpaceTransform UserTransform		= {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-	const AvatarState UserAvatarState		= AvatarState::Idle;
-	const csp::common::String UserAvatarId	= "MyCoolAvatar";
-	const AvatarPlayMode UserAvatarPlayMode = AvatarPlayMode::Default;
-
-	{
-
-		auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-		EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-		Connection	 = EnterResult.GetConnection();
-		EntitySystem = Connection->GetSpaceEntitySystem();
-
-		// Ensure we're in the first space
-		EXPECT_EQ(SpaceSystem->GetCurrentSpace().Id, Space.Id);
-
-		// Create Avatar
-		EntitySystem->SetEntityCreatedCallback(
-			[](SpaceEntity* Entity)
-			{
-			});
-
-		auto [Avatar] = AWAIT(EntitySystem, CreateAvatar, UserName, UserTransform, UserAvatarState, UserAvatarId, UserAvatarPlayMode);
-
-		// Create object to represent the portal
-		csp::common::String ObjectName = "Object 1";
-		SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-		auto [CreatedObject]		   = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-		// Create portal component
-		auto* PortalComponent = (PortalSpaceComponent*) CreatedObject->AddComponent(ComponentType::Portal);
-		PortalComponent->SetSpaceId(Space2.Id);
-
-		PortalSpaceID = PortalComponent->GetSpaceId();
-
-		AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-	}
-
-	/*
-		User would now interact with the portal
-	*/
-
-	{
-		auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-		EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-		Connection	 = EnterResult.GetConnection();
-		EntitySystem = Connection->GetSpaceEntitySystem();
-
-		// Create Avatar
-		EntitySystem->SetEntityCreatedCallback(
-			[](SpaceEntity* Entity)
-			{
-			});
-
-		auto [Avatar] = AWAIT(EntitySystem, CreateAvatar, UserName, UserTransform, UserAvatarState, UserAvatarId, UserAvatarPlayMode);
-
-		AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-	}
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-	DeleteSpace(SpaceSystem, Space2.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_PORTAL_THUMBNAIL_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, PortalThumbnailTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-	auto* AssetSystem	 = SystemsManager.GetAssetSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-
-	auto FilePath = std::filesystem::absolute("assets/OKO.png");
-
-	csp::systems::FileAssetDataSource Source;
-	Source.FilePath = FilePath.string().c_str();
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, Source, Space);
-
-	auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-	EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-	Connection	 = EnterResult.GetConnection();
-	EntitySystem = Connection->GetSpaceEntitySystem();
-
-	EntitySystem->SetEntityCreatedCallback(
-		[](SpaceEntity* Entity)
-		{
-		});
-
-	// Create object to represent the portal
-	csp::common::String ObjectName = "Object 1";
-	SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-	auto [CreatedObject]		   = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-	// Create portal component
-	auto* PortalComponent = (PortalSpaceComponent*) CreatedObject->AddComponent(ComponentType::Portal);
-
-	// Get Thumbnail
-	bool HasThumbailResult = false;
-
-	csp::systems::UriResultCallback Callback = [&HasThumbailResult](const csp::systems::UriResult& Result)
-	{
-		if (Result.GetResultCode() == csp::services::EResultCode::Success)
-		{
-			HasThumbailResult = true;
-			EXPECT_TRUE(Result.GetUri() != "");
-		}
-	};
-
-	PortalComponent->SetSpaceId(Space.Id);
-	PortalComponent->GetSpaceThumbnail(Callback);
-
-	auto Start		 = std::chrono::steady_clock::now();
-	auto Current	 = std::chrono::steady_clock::now();
-	int64_t TestTime = 0;
-
-	while (!HasThumbailResult && TestTime < 20)
-	{
-		std::this_thread::sleep_for(50ms);
-
-		Current	 = std::chrono::steady_clock::now();
-		TestTime = std::chrono::duration_cast<std::chrono::seconds>(Current - Start).count();
-	}
-
-	EXPECT_TRUE(HasThumbailResult);
-
-	AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
 
 #if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_DELETE_MULTIPLE_ENTITIES_TEST
 CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, DeleteMultipleEntitiesTest)
@@ -1970,1737 +1773,6 @@ CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, EntitySelectionTest)
 }
 #endif
 
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_CONVERSATION_COMPONENT_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, ConversationComponentTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	const char* TestSpaceName2		  = "OLY-UNITTEST-SPACE-REWIND-2";
-	const char* TestSpaceDescription2 = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	char UniqueSpaceName2[256];
-	SPRINTF(UniqueSpaceName2, "%s-%s", TestSpaceName2, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-	const auto UserDisplayName = GetFullProfileByUserId(UserSystem, UserId).DisplayName;
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	const csp::common::String UserName		= "Player 1";
-	const SpaceTransform UserTransform		= {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-	const AvatarState UserAvatarState		= AvatarState::Idle;
-	const csp::common::String UserAvatarId	= "MyCoolAvatar";
-	const AvatarPlayMode UserAvatarPlayMode = AvatarPlayMode::Default;
-
-	{
-
-		auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-		EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-		Connection	 = EnterResult.GetConnection();
-		EntitySystem = Connection->GetSpaceEntitySystem();
-
-		// Ensure we're in the first space
-		EXPECT_EQ(SpaceSystem->GetCurrentSpace().Id, Space.Id);
-
-		// Create Avatar
-		EntitySystem->SetEntityCreatedCallback(
-			[](SpaceEntity* Entity)
-			{
-			});
-
-		auto [Avatar] = AWAIT(EntitySystem, CreateAvatar, UserName, UserTransform, UserAvatarState, UserAvatarId, UserAvatarPlayMode);
-
-		// Create object to represent the conversation
-		csp::common::String ObjectName = "Object 1";
-		SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-		auto [CreatedObject]		   = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-		// Create conversation component
-		auto* ConversationComponent = (ConversationSpaceComponent*) CreatedObject->AddComponent(ComponentType::Conversation);
-
-		EXPECT_EQ(ConversationComponent->GetIsVisible(), true);
-		EXPECT_EQ(ConversationComponent->GetIsActive(), true);
-
-		ConversationComponent->SetIsActive(false);
-		ConversationComponent->SetIsVisible(false);
-
-		EXPECT_EQ(ConversationComponent->GetIsVisible(), false);
-		EXPECT_EQ(ConversationComponent->GetIsActive(), false);
-
-		SpaceTransform DefaultTransform = SpaceTransform();
-
-		EXPECT_EQ(ConversationComponent->GetPosition().X, DefaultTransform.Position.X);
-		EXPECT_EQ(ConversationComponent->GetPosition().Y, DefaultTransform.Position.Y);
-		EXPECT_EQ(ConversationComponent->GetPosition().Z, DefaultTransform.Position.Z);
-
-		csp::common::Vector3 NewPosition(1, 2, 3);
-		ConversationComponent->SetPosition(NewPosition);
-
-		EXPECT_EQ(ConversationComponent->GetPosition().X, NewPosition.X);
-		EXPECT_EQ(ConversationComponent->GetPosition().Y, NewPosition.Y);
-		EXPECT_EQ(ConversationComponent->GetPosition().Z, NewPosition.Z);
-
-		EXPECT_EQ(ConversationComponent->GetRotation().W, DefaultTransform.Rotation.W);
-		EXPECT_EQ(ConversationComponent->GetRotation().X, DefaultTransform.Rotation.X);
-		EXPECT_EQ(ConversationComponent->GetRotation().Y, DefaultTransform.Rotation.Y);
-		EXPECT_EQ(ConversationComponent->GetRotation().Z, DefaultTransform.Rotation.Z);
-
-		csp::common::Vector4 NewRotation(4, 5, 6, 7);
-		ConversationComponent->SetRotation(NewRotation);
-
-		EXPECT_EQ(ConversationComponent->GetRotation().W, NewRotation.W);
-		EXPECT_EQ(ConversationComponent->GetRotation().X, NewRotation.X);
-		EXPECT_EQ(ConversationComponent->GetRotation().Y, NewRotation.Y);
-		EXPECT_EQ(ConversationComponent->GetRotation().Z, NewRotation.Z);
-
-		EXPECT_EQ(ConversationComponent->GetTitle(), "");
-		EXPECT_EQ(ConversationComponent->GetDate(), "");
-		EXPECT_EQ(ConversationComponent->GetNumberOfReplies(), 0);
-
-		ConversationComponent->SetTitle("TestTitle");
-		ConversationComponent->SetDate("02-01-1972");
-		ConversationComponent->SetNumberOfReplies(2);
-
-		EXPECT_EQ(ConversationComponent->GetTitle(), "TestTitle");
-		EXPECT_EQ(ConversationComponent->GetDate(), "02-01-1972");
-		EXPECT_EQ(ConversationComponent->GetNumberOfReplies(), 2);
-
-		csp::common::String ConversationId;
-		csp::common::String MessageId;
-		{
-			auto [Result] = AWAIT(ConversationComponent, CreateConversation, "TestMessage");
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			EXPECT_TRUE(Result.GetValue() != "");
-			ConversationId = Result.GetValue();
-		}
-
-		{
-			auto [Result] = AWAIT_PRE(ConversationComponent, AddMessage, RequestPredicate, "Test");
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			MessageId = Result.GetMessageInfo().Id;
-			EXPECT_EQ(Result.GetMessageInfo().Edited, false);
-		}
-
-		{
-			auto [Result] = AWAIT(ConversationComponent, GetMessageInfo, MessageId);
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			EXPECT_EQ(Result.GetMessageInfo().Edited, false);
-		}
-
-		{
-			MessageInfo NewData = MessageInfo();
-			NewData.Message		= "NewTest";
-			auto [Result]		= AWAIT(ConversationComponent, SetMessageInfo, MessageId, NewData);
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			EXPECT_EQ(Result.GetMessageInfo().Edited, true);
-		}
-
-		{
-			auto [Result] = AWAIT(ConversationComponent, GetConversationInfo);
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			EXPECT_EQ(Result.GetConversationInfo().UserID, UserId);
-			EXPECT_EQ(Result.GetConversationInfo().UserDisplayName, UserDisplayName);
-			EXPECT_EQ(Result.GetConversationInfo().Message, "TestMessage");
-			EXPECT_FALSE(Result.GetConversationInfo().Edited);
-			EXPECT_FALSE(Result.GetConversationInfo().Resolved);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.X, DefaultTransform.Position.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.Y, DefaultTransform.Position.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.Z, DefaultTransform.Position.Z);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.W, DefaultTransform.Rotation.W);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.X, DefaultTransform.Rotation.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.Y, DefaultTransform.Rotation.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.Z, DefaultTransform.Rotation.Z);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.X, DefaultTransform.Scale.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.Y, DefaultTransform.Scale.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.Z, DefaultTransform.Scale.Z);
-		}
-
-		{
-			ConversationInfo NewData = ConversationInfo();
-			SpaceTransform CameraTransformValue
-				= SpaceTransform(csp::common::Vector3().One(), csp::common::Vector4().One(), csp::common::Vector3().One());
-			NewData.Resolved	   = true;
-			NewData.CameraPosition = CameraTransformValue;
-			NewData.Message		   = "TestMessage1";
-
-			auto [Result] = AWAIT(ConversationComponent, SetConversationInfo, NewData);
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			EXPECT_EQ(Result.GetConversationInfo().UserID, UserId);
-			EXPECT_EQ(Result.GetConversationInfo().UserDisplayName, UserDisplayName);
-			EXPECT_EQ(Result.GetConversationInfo().Message, "TestMessage1");
-			EXPECT_TRUE(Result.GetConversationInfo().Edited);
-			EXPECT_TRUE(Result.GetConversationInfo().Resolved);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.X, CameraTransformValue.Position.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.Y, CameraTransformValue.Position.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.Z, CameraTransformValue.Position.Z);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.W, CameraTransformValue.Rotation.W);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.X, CameraTransformValue.Rotation.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.Y, CameraTransformValue.Rotation.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.Z, CameraTransformValue.Rotation.Z);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.X, CameraTransformValue.Scale.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.Y, CameraTransformValue.Scale.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.Z, CameraTransformValue.Scale.Z);
-			EXPECT_EQ(Result.GetConversationInfo().Message, "TestMessage1");
-		}
-
-		auto TestMessage = "test123";
-		Connection->ListenNetworkEvent("ConversationSystem:NewMessage",
-									   [=](bool ok, csp::common::Array<ReplicatedValue> Data)
-									   {
-										   EXPECT_TRUE(ok);
-										   ConversationId == Data[0].GetString();
-										   std::cerr << "Test Event Received " << ok << std::endl;
-									   });
-
-		{
-			auto [Result] = AWAIT_PRE(ConversationComponent, AddMessage, RequestPredicate, TestMessage);
-
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			MessageId = Result.GetMessageInfo().Id;
-		}
-		{
-			auto [Result] = AWAIT(ConversationComponent, GetAllMessages);
-
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			EXPECT_EQ(Result.GetTotalCount(), 2);
-
-			EXPECT_EQ(Result.GetMessages()[0].Id, MessageId);
-		}
-		{
-			auto [Result] = AWAIT(ConversationComponent, GetMessage, MessageId);
-
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			EXPECT_EQ(Result.GetMessageInfo().Id, MessageId);
-		}
-
-		{
-			auto [Result] = AWAIT(ConversationComponent, DeleteMessage, MessageId);
-
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-		}
-
-		{
-			auto [Result] = AWAIT(ConversationComponent, DeleteConversation);
-
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-		}
-
-		AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-	}
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_CONVERSATION_COMPONENT_MOVE_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, ConversationComponentMoveTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	const char* TestSpaceName2		  = "OLY-UNITTEST-SPACE-REWIND-2";
-	const char* TestSpaceDescription2 = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-	const auto UserDisplayName = GetFullProfileByUserId(UserSystem, UserId).DisplayName;
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	{
-
-		auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-		EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-		Connection	 = EnterResult.GetConnection();
-		EntitySystem = Connection->GetSpaceEntitySystem();
-
-		// Ensure we're in the first space
-		EXPECT_EQ(SpaceSystem->GetCurrentSpace().Id, Space.Id);
-
-		// Create Avatar
-		EntitySystem->SetEntityCreatedCallback(
-			[](SpaceEntity* Entity)
-			{
-			});
-
-		csp::common::String ObjectName1 = "Object 1";
-		csp::common::String ObjectName2 = "Object 2";
-
-		SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-
-		auto [CreatedObject1] = AWAIT(EntitySystem, CreateObject, ObjectName1, ObjectTransform);
-		auto [CreatedObject2] = AWAIT(EntitySystem, CreateObject, ObjectName2, ObjectTransform);
-
-		// Create conversation component
-		auto* ConversationComponent1 = (ConversationSpaceComponent*) CreatedObject1->AddComponent(ComponentType::Conversation);
-		auto* ConversationComponent2 = (ConversationSpaceComponent*) CreatedObject2->AddComponent(ComponentType::Conversation);
-
-
-		csp::common::String ConversationId;
-		csp::common::String MessageId;
-		{
-			auto [Result] = AWAIT(ConversationComponent1, CreateConversation, "TestMessage");
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			EXPECT_TRUE(Result.GetValue() != "");
-			ConversationId = Result.GetValue();
-		}
-
-		SpaceTransform DefaultTransform = SpaceTransform();
-
-		{
-			auto [Result] = AWAIT(ConversationComponent1, GetConversationInfo);
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			EXPECT_EQ(Result.GetConversationInfo().UserID, UserId);
-			EXPECT_EQ(Result.GetConversationInfo().UserDisplayName, UserDisplayName);
-			EXPECT_EQ(Result.GetConversationInfo().Message, "TestMessage");
-			EXPECT_FALSE(Result.GetConversationInfo().Edited);
-			EXPECT_FALSE(Result.GetConversationInfo().Resolved);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.X, DefaultTransform.Position.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.Y, DefaultTransform.Position.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.Z, DefaultTransform.Position.Z);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.W, DefaultTransform.Rotation.W);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.X, DefaultTransform.Rotation.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.Y, DefaultTransform.Rotation.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.Z, DefaultTransform.Rotation.Z);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.X, DefaultTransform.Scale.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.Y, DefaultTransform.Scale.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.Z, DefaultTransform.Scale.Z);
-		}
-
-		{
-			auto [Result] = AWAIT(ConversationComponent2, GetConversationInfo);
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Failed);
-		}
-
-		{
-			auto Result = ConversationComponent2->MoveConversationFromComponent(*ConversationComponent1);
-			EXPECT_TRUE(Result);
-		}
-
-		{
-			auto [Result] = AWAIT(ConversationComponent1, GetConversationInfo);
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Failed);
-		}
-
-		{
-			auto [Result] = AWAIT(ConversationComponent2, GetConversationInfo);
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-			EXPECT_EQ(Result.GetConversationInfo().UserID, UserId);
-			EXPECT_EQ(Result.GetConversationInfo().UserDisplayName, UserDisplayName);
-			EXPECT_EQ(Result.GetConversationInfo().Message, "TestMessage");
-			EXPECT_FALSE(Result.GetConversationInfo().Edited);
-			EXPECT_FALSE(Result.GetConversationInfo().Resolved);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.X, DefaultTransform.Position.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.Y, DefaultTransform.Position.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Position.Z, DefaultTransform.Position.Z);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.W, DefaultTransform.Rotation.W);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.X, DefaultTransform.Rotation.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.Y, DefaultTransform.Rotation.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Rotation.Z, DefaultTransform.Rotation.Z);
-
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.X, DefaultTransform.Scale.X);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.Y, DefaultTransform.Scale.Y);
-			EXPECT_EQ(Result.GetConversationInfo().CameraPosition.Scale.Z, DefaultTransform.Scale.Z);
-		}
-
-		{
-			auto [Result] = AWAIT(ConversationComponent2, DeleteConversation);
-			EXPECT_EQ(Result.GetResultCode(), csp::services::EResultCode::Success);
-		}
-
-		AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-	}
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_AUDIO_COMPONENT_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, AudioComponentTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-	EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-	Connection	 = EnterResult.GetConnection();
-	EntitySystem = Connection->GetSpaceEntitySystem();
-
-	EntitySystem->SetEntityCreatedCallback(
-		[](SpaceEntity* Entity)
-		{
-		});
-
-	// Create object to represent the audio
-	csp::common::String ObjectName = "Object 1";
-	SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-	auto [CreatedObject]		   = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-	// Create audio component
-	auto* AudioComponent = static_cast<AudioSpaceComponent*>(CreatedObject->AddComponent(ComponentType::Audio));
-
-	// Ensure defaults are set
-	EXPECT_EQ(AudioComponent->GetPosition(), csp::common::Vector3::Zero());
-	EXPECT_EQ(AudioComponent->GetPlaybackState(), AudioPlaybackState::Reset);
-	EXPECT_EQ(AudioComponent->GetAudioType(), AudioType::Global);
-	EXPECT_EQ(AudioComponent->GetAudioAssetId(), "");
-	EXPECT_EQ(AudioComponent->GetAssetCollectionId(), "");
-	EXPECT_EQ(AudioComponent->GetAttenuationRadius(), 10.f);
-	EXPECT_EQ(AudioComponent->GetIsLoopPlayback(), false);
-	EXPECT_EQ(AudioComponent->GetTimeSincePlay(), 0.f);
-	EXPECT_EQ(AudioComponent->GetVolume(), 1.f);
-	EXPECT_EQ(AudioComponent->GetIsEnabled(), true);
-
-	// Set new values
-	csp::common::String AssetId			  = "TEST_ASSET_ID";
-	csp::common::String AssetCollectionId = "TEST_COLLECTION_ID";
-
-	AudioComponent->SetPosition(csp::common::Vector3::One());
-	AudioComponent->SetPlaybackState(AudioPlaybackState::Play);
-	AudioComponent->SetAudioType(AudioType::Spatial);
-	AudioComponent->SetAudioAssetId(AssetId);
-	AudioComponent->SetAssetCollectionId(AssetCollectionId);
-	AudioComponent->SetAttenuationRadius(100.f);
-	AudioComponent->SetIsLoopPlayback(true);
-	AudioComponent->SetTimeSincePlay(1.f);
-	AudioComponent->SetVolume(0.5f);
-	AudioComponent->SetIsEnabled(false);
-
-	// Ensure values are set correctly
-	EXPECT_EQ(AudioComponent->GetPosition(), csp::common::Vector3::One());
-	EXPECT_EQ(AudioComponent->GetPlaybackState(), AudioPlaybackState::Play);
-	EXPECT_EQ(AudioComponent->GetAudioType(), AudioType::Spatial);
-	EXPECT_EQ(AudioComponent->GetAudioAssetId(), AssetId);
-	EXPECT_EQ(AudioComponent->GetAssetCollectionId(), AssetCollectionId);
-	EXPECT_EQ(AudioComponent->GetAttenuationRadius(), 100.f);
-	EXPECT_EQ(AudioComponent->GetIsLoopPlayback(), true);
-	EXPECT_EQ(AudioComponent->GetTimeSincePlay(), 1.f);
-	EXPECT_EQ(AudioComponent->GetVolume(), 0.5f);
-	EXPECT_EQ(AudioComponent->GetIsEnabled(), false);
-
-	// Test invalid volume values
-	AudioComponent->SetVolume(1.5f);
-	EXPECT_EQ(AudioComponent->GetVolume(), 0.5f);
-	AudioComponent->SetVolume(-2.5f);
-	EXPECT_EQ(AudioComponent->GetVolume(), 0.5f);
-
-	// Test boundary volume values
-	AudioComponent->SetVolume(1.f);
-	EXPECT_EQ(AudioComponent->GetVolume(), 1.f);
-	AudioComponent->SetVolume(0.f);
-	EXPECT_EQ(AudioComponent->GetVolume(), 0.f);
-
-	AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_VIDEO_PLAYER_COMPONENT_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, VideoPlayerComponentTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-	EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-	Connection	 = EnterResult.GetConnection();
-	EntitySystem = Connection->GetSpaceEntitySystem();
-
-	EntitySystem->SetEntityCreatedCallback(
-		[](SpaceEntity* Entity)
-		{
-		});
-
-	// Create object to represent the audio
-	csp::common::String ObjectName = "Object 1";
-	SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-	auto [CreatedObject]		   = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-	// Create audio component
-	auto* VideoComponent = static_cast<VideoPlayerSpaceComponent*>(CreatedObject->AddComponent(ComponentType::VideoPlayer));
-
-	// Ensure defaults are set
-	EXPECT_EQ(VideoComponent->GetPosition(), csp::common::Vector3::Zero());
-	EXPECT_EQ(VideoComponent->GetPlaybackState(), VideoPlayerPlaybackState::Reset);
-	EXPECT_EQ(VideoComponent->GetVideoAssetURL(), "");
-	EXPECT_EQ(VideoComponent->GetAssetCollectionId(), "");
-	EXPECT_EQ(VideoComponent->GetAttenuationRadius(), 10.f);
-	EXPECT_EQ(VideoComponent->GetIsLoopPlayback(), false);
-	EXPECT_EQ(VideoComponent->GetTimeSincePlay(), 0.f);
-	EXPECT_EQ(VideoComponent->GetIsStateShared(), false);
-	EXPECT_EQ(VideoComponent->GetIsAutoPlay(), false);
-	EXPECT_EQ(VideoComponent->GetIsAutoResize(), false);
-	EXPECT_EQ(VideoComponent->GetCurrentPlayheadPosition(), 0.0f);
-	EXPECT_EQ(VideoComponent->GetVideoPlayerSourceType(), VideoPlayerSourceType::AssetSource);
-	EXPECT_EQ(VideoComponent->GetIsVisible(), true);
-	EXPECT_EQ(VideoComponent->GetMeshComponentId(), 0);
-
-	auto* ModelComponent = static_cast<VideoPlayerSpaceComponent*>(CreatedObject->AddComponent(ComponentType::AnimatedModel));
-
-	CreatedObject->QueueUpdate();
-	EntitySystem->ProcessPendingEntityOperations();
-
-	// Set new values
-	csp::common::String AssetId			  = "TEST_ASSET_ID";
-	csp::common::String AssetCollectionId = "TEST_COLLECTION_ID";
-
-	VideoComponent->SetPosition(csp::common::Vector3::One());
-	VideoComponent->SetPlaybackState(VideoPlayerPlaybackState::Play);
-	VideoComponent->SetVideoAssetURL("http://youtube.com/avideo");
-	VideoComponent->SetAssetCollectionId(AssetId);
-	VideoComponent->SetAttenuationRadius(100.f);
-	VideoComponent->SetIsLoopPlayback(true);
-	VideoComponent->SetTimeSincePlay(1.f);
-	VideoComponent->SetIsStateShared(true);
-	VideoComponent->SetIsAutoPlay(true);
-	VideoComponent->SetIsAutoResize(true);
-	VideoComponent->SetCurrentPlayheadPosition(1.0f);
-	VideoComponent->SetVideoPlayerSourceType(VideoPlayerSourceType::URLSource);
-	VideoComponent->SetIsVisible(false);
-	VideoComponent->SetMeshComponentId(ModelComponent->GetId());
-
-	// Ensure values are set correctly
-	EXPECT_EQ(VideoComponent->GetPosition(), csp::common::Vector3::One());
-	EXPECT_EQ(VideoComponent->GetPlaybackState(), VideoPlayerPlaybackState::Play);
-	EXPECT_EQ(VideoComponent->GetVideoAssetURL(), "http://youtube.com/avideo");
-	EXPECT_EQ(VideoComponent->GetAssetCollectionId(), AssetId);
-	EXPECT_EQ(VideoComponent->GetAttenuationRadius(), 100.f);
-	EXPECT_EQ(VideoComponent->GetIsLoopPlayback(), true);
-	EXPECT_EQ(VideoComponent->GetTimeSincePlay(), 1.f);
-	EXPECT_EQ(VideoComponent->GetIsStateShared(), true);
-	EXPECT_EQ(VideoComponent->GetIsAutoPlay(), true);
-	EXPECT_EQ(VideoComponent->GetIsAutoResize(), true);
-	EXPECT_EQ(VideoComponent->GetCurrentPlayheadPosition(), 1.0f);
-	EXPECT_EQ(VideoComponent->GetVideoPlayerSourceType(), VideoPlayerSourceType::URLSource);
-	EXPECT_EQ(VideoComponent->GetIsVisible(), false);
-	EXPECT_EQ(VideoComponent->GetMeshComponentId(), ModelComponent->GetId());
-
-	AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_COLLISION_COMPONENT_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, CollisionComponentTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-	EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-	Connection	 = EnterResult.GetConnection();
-	EntitySystem = Connection->GetSpaceEntitySystem();
-
-	EntitySystem->SetEntityCreatedCallback(
-		[](SpaceEntity* Entity)
-		{
-		});
-
-	// Create object to represent the audio
-	csp::common::String ObjectName = "Object 1";
-	SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-	auto [CreatedObject]		   = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-	// Create collision component
-	auto* CollisionComponent = static_cast<CollisionSpaceComponent*>(CreatedObject->AddComponent(ComponentType::Collision));
-
-	// Ensure defaults are set
-	EXPECT_EQ(CollisionComponent->GetPosition(), csp::common::Vector3::Zero());
-	EXPECT_EQ(CollisionComponent->GetRotation(), csp::common::Vector4(0, 0, 0, 1));
-	EXPECT_EQ(CollisionComponent->GetScale(), csp::common::Vector3::One());
-	EXPECT_EQ(CollisionComponent->GetUnscaledBoundingBoxMin(), csp::common::Vector3(-0.5, -0.5, -0.5));
-	EXPECT_EQ(CollisionComponent->GetUnscaledBoundingBoxMax(), csp::common::Vector3(0.5, 0.5, 0.5));
-	EXPECT_EQ(CollisionComponent->GetScaledBoundingBoxMin(), csp::common::Vector3(-0.5, -0.5, -0.5));
-	EXPECT_EQ(CollisionComponent->GetScaledBoundingBoxMax(), csp::common::Vector3(0.5, 0.5, 0.5));
-	EXPECT_EQ(CollisionComponent->GetCollisionMode(), csp::multiplayer::CollisionMode::Collision);
-	EXPECT_EQ(CollisionComponent->GetCollisionShape(), csp::multiplayer::CollisionShape::Box);
-	EXPECT_EQ(CollisionComponent->GetCollisionAssetId(), "");
-	EXPECT_EQ(CollisionComponent->GetAssetCollectionId(), "");
-
-	// Set new values
-
-	CollisionComponent->SetPosition(csp::common::Vector3::One());
-	CollisionComponent->SetScale(csp::common::Vector3(2, 2, 2));
-	CollisionComponent->SetCollisionMode(csp::multiplayer::CollisionMode::Trigger);
-	CollisionComponent->SetCollisionShape(csp::multiplayer::CollisionShape::Mesh);
-	CollisionComponent->SetCollisionAssetId("TestAssetID");
-	CollisionComponent->SetAssetCollectionId("TestAssetCollectionID");
-
-	// Ensure values are set correctly
-	EXPECT_EQ(CollisionComponent->GetPosition(), csp::common::Vector3::One());
-	EXPECT_EQ(CollisionComponent->GetScale(), csp::common::Vector3(2, 2, 2));
-	EXPECT_EQ(CollisionComponent->GetUnscaledBoundingBoxMin(), csp::common::Vector3(-0.5, -0.5, -0.5));
-	EXPECT_EQ(CollisionComponent->GetUnscaledBoundingBoxMax(), csp::common::Vector3(0.5, 0.5, 0.5));
-	EXPECT_EQ(CollisionComponent->GetScaledBoundingBoxMin(), csp::common::Vector3(-1, -1, -1));
-	EXPECT_EQ(CollisionComponent->GetScaledBoundingBoxMax(), csp::common::Vector3(1, 1, 1));
-	EXPECT_EQ(CollisionComponent->GetCollisionMode(), csp::multiplayer::CollisionMode::Trigger);
-	EXPECT_EQ(CollisionComponent->GetCollisionShape(), csp::multiplayer::CollisionShape::Mesh);
-	EXPECT_EQ(CollisionComponent->GetCollisionAssetId(), "TestAssetID");
-	EXPECT_EQ(CollisionComponent->GetAssetCollectionId(), "TestAssetCollectionID");
-
-	const float DefaultSphereRadius		 = CollisionSpaceComponent::GetDefaultSphereRadius();
-	const float DefaultCapsuleHalfWidth	 = CollisionSpaceComponent::GetDefaultCapsuleHalfWidth();
-	const float DefaultCapsuleHalfHeight = CollisionSpaceComponent::GetDefaultCapsuleHalfHeight();
-
-	EXPECT_EQ(DefaultSphereRadius, 0.5f);
-	EXPECT_EQ(DefaultCapsuleHalfWidth, 0.5f);
-	EXPECT_EQ(DefaultCapsuleHalfHeight, 1.0f);
-
-	AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_USE_SPLINE_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, UseSplineTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	const csp::common::String UserName		= "Player 1";
-	const SpaceTransform UserTransform		= {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-	const AvatarState UserAvatarState		= AvatarState::Idle;
-	const csp::common::String UserAvatarId	= "MyCoolAvatar";
-	const AvatarPlayMode UserAvatarPlayMode = AvatarPlayMode::Default;
-
-	{
-
-		auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-		EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-		Connection	 = EnterResult.GetConnection();
-		EntitySystem = Connection->GetSpaceEntitySystem();
-
-		EntitySystem->SetEntityCreatedCallback(
-			[](SpaceEntity* Entity)
-			{
-			});
-
-		// Ensure we're in the first space
-		EXPECT_EQ(SpaceSystem->GetCurrentSpace().Id, Space.Id);
-
-		// Create object to represent the spline
-		csp::common::String ObjectName = "Object 1";
-		SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-		auto [CreatedObject]		   = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-		// Create spline component
-		auto* SplineComponent							  = (SplineSpaceComponent*) CreatedObject->AddComponent(ComponentType::Spline);
-		csp::common::List<csp::common::Vector3> WayPoints = {{0, 0, 0}, {0, 1000, 0}, {0, 2000, 0}, {0, 3000, 0}, {0, 4000, 0}, {0, 5000, 0}};
-
-		{
-			auto Result = SplineComponent->GetWaypoints();
-
-			EXPECT_EQ(Result.Size(), 0);
-		}
-
-		{
-			auto Result = SplineComponent->GetLocationAlongSpline(1);
-			EXPECT_EQ(Result.X, 0);
-			EXPECT_EQ(Result.Y, 0);
-			EXPECT_EQ(Result.Z, 0);
-		}
-
-		{
-			SplineComponent->SetWaypoints(WayPoints);
-
-			auto Result = SplineComponent->GetWaypoints();
-
-			EXPECT_EQ(Result.Size(), WayPoints.Size());
-
-			// expect final waypoint to be the same
-			EXPECT_EQ(Result[0], WayPoints[0]);
-		}
-
-		{
-			// Calculated cubic interpolate spline
-			auto Result = SplineComponent->GetLocationAlongSpline(1);
-
-			// expect final waypoint to be the same
-			EXPECT_EQ(Result, WayPoints[WayPoints.Size() - 1]);
-		}
-
-		AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-	}
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_LIGHT_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, LightComponentFieldsTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-	auto* AssetSystem	 = SystemsManager.GetAssetSystem();
-
-	const char* TestSpaceName			= "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription	= "OLY-UNITTEST-SPACEDESC-REWIND";
-	const char* TestAssetCollectionName = "OLY-UNITTEST-ASSETCOLLECTION-REWIND";
-	const char* TestAssetName			= "OLY-UNITTEST-ASSET-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	csp::common::String UserId;
-
-	// Log in
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-	EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-	Connection	 = EnterResult.GetConnection();
-	EntitySystem = Connection->GetSpaceEntitySystem();
-
-	EntitySystem->SetEntityCreatedCallback(
-		[](SpaceEntity* Entity)
-		{
-		});
-
-	bool AssetDetailBlobChangedCallbackCalled = false;
-	csp::common::String CallbackAssetId;
-
-	const csp::common::String ObjectName = "Object 1";
-	SpaceTransform ObjectTransform		 = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-
-	auto [Object] = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-	const csp::common::String ModelAssetId = "NotARealId";
-
-	auto* LightSpaceComponentInstance = (LightSpaceComponent*) Object->AddComponent(ComponentType::Light);
-
-	// Process component creation
-	Object->QueueUpdate();
-	EntitySystem->ProcessPendingEntityOperations();
-
-	// Check component was created
-	auto& Components = *Object->GetComponents();
-	ASSERT_EQ(Components.Size(), 1);
-
-	char UniqueAssetCollectionName[256];
-	SPRINTF(UniqueAssetCollectionName, "%s-%s", TestAssetCollectionName, GetUniqueHexString().c_str());
-
-	char UniqueAssetName[256];
-	SPRINTF(UniqueAssetName, "%s-%s", TestAssetName, GetUniqueHexString().c_str());
-
-	// Create asset collection
-	csp::systems::AssetCollection AssetCollection;
-	CreateAssetCollection(AssetSystem, Space.Id, nullptr, UniqueAssetCollectionName, nullptr, nullptr, AssetCollection);
-
-	// Create asset
-	csp::systems::Asset Asset;
-	CreateAsset(AssetSystem, AssetCollection, UniqueAssetName, nullptr, nullptr, Asset);
-	Asset.FileName = "OKO.png";
-	Asset.Name	   = "OKO";
-	Asset.Type	   = csp::systems::EAssetType::IMAGE;
-
-	auto UploadFilePath		 = std::filesystem::absolute("assets/OKO.png");
-	FILE* UploadFile		 = fopen(UploadFilePath.string().c_str(), "rb");
-	uintmax_t UploadFileSize = std::filesystem::file_size(UploadFilePath);
-	auto* UploadFileData	 = new unsigned char[UploadFileSize];
-	fread(UploadFileData, UploadFileSize, 1, UploadFile);
-	fclose(UploadFile);
-
-	csp::systems::BufferAssetDataSource BufferSource;
-	BufferSource.Buffer		  = UploadFileData;
-	BufferSource.BufferLength = UploadFileSize;
-
-	BufferSource.SetMimeType("image/png");
-
-	printf("Uploading asset data...\n");
-
-	// Upload data
-	UploadAssetData(AssetSystem, AssetCollection, Asset, BufferSource, Asset.Uri);
-
-	delete[] UploadFileData;
-
-	EXPECT_EQ(LightSpaceComponentInstance->GetLightCookieType(), LightCookieType::NoCookie);
-	EXPECT_EQ(LightSpaceComponentInstance->GetLightType(), LightType::Point);
-	EXPECT_EQ(LightSpaceComponentInstance->GetLightShadowType(), LightShadowType::None);
-
-	// test values
-	const float InnerConeAngle = 10.0f;
-	const float OuterConeAngle = 20.0f;
-	const float Range		   = 120.0f;
-	const float Intensity	   = 1000.0f;
-
-	LightSpaceComponentInstance->SetLightCookieType(LightCookieType::ImageCookie);
-	LightSpaceComponentInstance->SetLightCookieAssetCollectionId(Asset.AssetCollectionId);
-	LightSpaceComponentInstance->SetLightCookieAssetId(Asset.Id);
-	LightSpaceComponentInstance->SetLightType(LightType::Spot);
-	LightSpaceComponentInstance->SetLightShadowType(LightShadowType::Realtime);
-	LightSpaceComponentInstance->SetInnerConeAngle(InnerConeAngle);
-	LightSpaceComponentInstance->SetOuterConeAngle(OuterConeAngle);
-	LightSpaceComponentInstance->SetRange(Range);
-	LightSpaceComponentInstance->SetIntensity(Intensity);
-
-	auto LightSpaceComponentKey				= LightSpaceComponentInstance->GetId();
-	auto* StoredLightSpaceComponentInstance = (LightSpaceComponent*) Object->GetComponent(LightSpaceComponentKey);
-
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetLightCookieType(), LightCookieType::ImageCookie);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetLightCookieAssetCollectionId(), Asset.AssetCollectionId);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetLightCookieAssetId(), Asset.Id);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetLightType(), LightType::Spot);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetLightShadowType(), LightShadowType::Realtime);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetInnerConeAngle(), InnerConeAngle);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetOuterConeAngle(), OuterConeAngle);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetRange(), Range);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetIntensity(), Intensity);
-
-	AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-
-	// Delete asset
-	DeleteAsset(AssetSystem, AssetCollection, Asset);
-
-	// Delete asset collection
-	DeleteAssetCollection(AssetSystem, AssetCollection);
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_CUSTOM_COMPONENT_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, CustomComponentTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	const char* TestSpaceName2		  = "OLY-UNITTEST-SPACE-REWIND-2";
-	const char* TestSpaceDescription2 = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	const csp::common::String ObjectName		= "Object 1";
-	const csp::common::String ApplicationOrigin = "Application Origin 1";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	char UniqueSpaceName2[256];
-	SPRINTF(UniqueSpaceName2, "%s-%s", TestSpaceName2, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	{
-
-		auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-		EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-		Connection	 = EnterResult.GetConnection();
-		EntitySystem = Connection->GetSpaceEntitySystem();
-
-		EntitySystem->SetEntityCreatedCallback(
-			[](SpaceEntity* Entity)
-			{
-			});
-
-		// Create object to represent the Custom fields
-		SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-		auto [CreatedObject]		   = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-		// Create custom component
-		auto* CustomComponent = (CustomSpaceComponent*) CreatedObject->AddComponent(ComponentType::Custom);
-
-		EXPECT_EQ(CustomComponent->GetCustomPropertyKeys().Size(), 0);
-
-		// Spectify the application origin and verify
-		CustomComponent->SetApplicationOrigin(ApplicationOrigin);
-
-		EXPECT_EQ(CustomComponent->GetApplicationOrigin(), ApplicationOrigin);
-
-		// Vector Check
-		{
-			CustomComponent->SetCustomProperty("Vector3", ReplicatedValue({10, 10, 10}));
-			EXPECT_EQ(CustomComponent->GetCustomProperty("Vector3").GetVector3(), csp::common::Vector3({10, 10, 10}));
-
-			CustomComponent->SetCustomProperty("Vector4", ReplicatedValue({10, 10, 10, 10}));
-			EXPECT_EQ(CustomComponent->GetCustomProperty("Vector4").GetVector4(), csp::common::Vector4({10, 10, 10, 10}));
-		}
-
-		// String Check
-		{
-			CustomComponent->SetCustomProperty("String", ReplicatedValue("OKO"));
-			EXPECT_EQ(CustomComponent->GetCustomProperty("String").GetString(), "OKO");
-		}
-
-		// Boolean Check
-		{
-			CustomComponent->SetCustomProperty("Boolean", ReplicatedValue(true));
-			EXPECT_EQ(CustomComponent->GetCustomProperty("Boolean").GetBool(), true);
-		}
-
-		// Integer Check
-		{
-			CustomComponent->SetCustomProperty("Integer", ReplicatedValue(int64_t(1)));
-			EXPECT_EQ(CustomComponent->GetCustomProperty("Integer").GetInt(), int64_t(1));
-		}
-
-		// Float Check
-		{
-			CustomComponent->SetCustomProperty("Float", ReplicatedValue(1.00f));
-			EXPECT_EQ(CustomComponent->GetCustomProperty("Float").GetFloat(), 1.00f);
-		}
-
-		// Has Key Check
-		{
-			EXPECT_EQ(CustomComponent->HasCustomProperty("Boolean"), true);
-			EXPECT_EQ(CustomComponent->HasCustomProperty("BooleanFalse"), false);
-		}
-
-		// Key Size
-		{
-			// Custom properties including application origin
-			EXPECT_EQ(CustomComponent->GetNumProperties(), 7);
-		}
-
-		// Remove Key
-		{
-			CustomComponent->RemoveCustomProperty("Boolean");
-
-			// Custom properties including application origin
-			EXPECT_EQ(CustomComponent->GetNumProperties(), 6);
-		}
-
-		// List Check
-		{
-			auto Keys = CustomComponent->GetCustomPropertyKeys();
-			EXPECT_EQ(Keys.Size(), 5);
-		}
-
-		// Queue update process before exiting space
-		EntitySystem->QueueEntityUpdate(CreatedObject);
-		EntitySystem->ProcessPendingEntityOperations();
-
-		AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-	}
-
-	// Re-Enter space and verify contents
-	{
-		// Reload the space and verify the contents match
-		auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-		EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-		Connection	 = EnterResult.GetConnection();
-		EntitySystem = Connection->GetSpaceEntitySystem();
-
-		// Retrieve all entities
-		auto GotAllEntities = false;
-		SpaceEntity* LoadedObject;
-
-		EntitySystem->SetEntityCreatedCallback(
-			[&](SpaceEntity* Entity)
-			{
-				if (Entity->GetName() == ObjectName)
-				{
-					GotAllEntities = true;
-					LoadedObject   = Entity;
-				}
-			});
-
-		// Wait until loaded
-		while (!GotAllEntities)
-		{
-			std::this_thread::sleep_for(100ms);
-		}
-
-		const auto& Components = *LoadedObject->GetComponents();
-		EXPECT_EQ(Components.Size(), 1);
-
-		// Retreive the custom component
-		auto LoadedComponent = Components[0];
-
-		// Verify the component type
-		EXPECT_EQ(LoadedComponent->GetComponentType(), ComponentType::Custom);
-
-		// Verify the application
-		auto* CustomComponent = (CustomSpaceComponent*) LoadedComponent;
-		EXPECT_EQ(CustomComponent->GetApplicationOrigin(), ApplicationOrigin);
-
-		// List Check
-		{
-			auto Keys = CustomComponent->GetCustomPropertyKeys();
-			EXPECT_EQ(Keys.Size(), 5);
-
-			// Vector Check
-			{
-				EXPECT_EQ(CustomComponent->GetCustomProperty("Vector3").GetVector3(), csp::common::Vector3({10, 10, 10}));
-
-				EXPECT_EQ(CustomComponent->GetCustomProperty("Vector4").GetVector4(), csp::common::Vector4({10, 10, 10, 10}));
-			}
-
-			// String Check
-			{
-				EXPECT_EQ(CustomComponent->GetCustomProperty("String").GetString(), "OKO");
-			}
-
-			// Integer Check
-			{
-				EXPECT_EQ(CustomComponent->GetCustomProperty("Integer").GetInt(), int64_t(1));
-			}
-
-			// Float Check
-			{
-				EXPECT_EQ(CustomComponent->GetCustomProperty("Float").GetFloat(), 1.00f);
-			}
-
-			// Has Missing Key Check
-			{
-				EXPECT_EQ(CustomComponent->HasCustomProperty("Boolean"), false);
-			}
-		}
-
-		AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-	}
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_EXTERNAL_LINK_COMPONENT_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, ExternalLinkComponentTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	const char* TestSpaceName2		  = "OLY-UNITTEST-SPACE-REWIND-2";
-	const char* TestSpaceDescription2 = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	{
-
-		auto* Connection   = new csp::multiplayer::MultiplayerConnection(Space.Id);
-		auto* EntitySystem = Connection->GetSpaceEntitySystem();
-
-		EntitySystem->SetEntityCreatedCallback(
-			[](SpaceEntity* Entity)
-			{
-			});
-
-		// Connect to the SignalR server
-		auto [Ok] = AWAIT(Connection, Connect);
-
-		EXPECT_TRUE(Ok);
-
-		std::tie(Ok) = AWAIT(Connection, InitialiseConnection);
-
-		EXPECT_TRUE(Ok);
-
-		csp::common::String ObjectName = "Object 1";
-		SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-		auto [CreatedObject]		   = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-		// Create custom component
-		auto* ExternalLinkComponent = (ExternalLinkSpaceComponent*) CreatedObject->AddComponent(ComponentType::ExternalLink);
-
-		const csp::common::String ExternalLinkName = "MyExternalLink";
-		ExternalLinkComponent->SetName(ExternalLinkName);
-		EXPECT_EQ(ExternalLinkComponent->GetName(), ExternalLinkName);
-
-		const csp::common::String ExternalLinkUrl = "https://oko.live";
-		ExternalLinkComponent->SetLinkUrl(ExternalLinkUrl);
-		EXPECT_EQ(ExternalLinkComponent->GetLinkUrl(), ExternalLinkUrl);
-
-		const csp::common::Vector3 Position(123.0f, 456.0f, 789.0f);
-		ExternalLinkComponent->SetPosition(Position);
-		EXPECT_EQ(ExternalLinkComponent->GetPosition(), Position);
-
-		const csp::common::Vector4 Rotation(1.0f, 2.0f, 3.0f, 4.0f);
-		ExternalLinkComponent->SetRotation(Rotation);
-		EXPECT_EQ(ExternalLinkComponent->GetRotation(), Rotation);
-
-		const csp::common::Vector3 Scale(123.0f, 456.0f, 789.0f);
-		ExternalLinkComponent->SetScale(Scale);
-		EXPECT_EQ(ExternalLinkComponent->GetScale(), Scale);
-
-		const csp::common::String DisplayText = "A great link";
-		ExternalLinkComponent->SetDisplayText(DisplayText);
-		EXPECT_EQ(ExternalLinkComponent->GetDisplayText(), DisplayText);
-
-		bool IsEnabled = false;
-		ExternalLinkComponent->SetIsEnabled(IsEnabled);
-		EXPECT_EQ(ExternalLinkComponent->GetIsEnabled(), IsEnabled);
-
-		bool IsVisible = false;
-		ExternalLinkComponent->SetIsVisible(IsVisible);
-		EXPECT_EQ(ExternalLinkComponent->GetIsVisible(), IsVisible);
-
-		bool IsARVisible = false;
-		ExternalLinkComponent->SetIsARVisible(IsARVisible);
-		EXPECT_EQ(ExternalLinkComponent->GetIsARVisible(), IsARVisible);
-
-		// Disconnect from the SignalR server
-		std::tie(Ok) = AWAIT(Connection, Disconnect);
-
-		EXPECT_TRUE(Ok);
-
-		// Delete MultiplayerConnection
-		delete Connection;
-	}
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_ACTIONHANDLER_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, ActionHandlerTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-	auto* AssetSystem	 = SystemsManager.GetAssetSystem();
-
-	const char* TestSpaceName			= "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription	= "OLY-UNITTEST-SPACEDESC-REWIND";
-	const char* TestAssetCollectionName = "OLY-UNITTEST-ASSETCOLLECTION-REWIND";
-	const char* TestAssetName			= "OLY-UNITTEST-ASSET-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	csp::common::String UserId;
-
-	// Log in
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-	EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-	Connection	 = EnterResult.GetConnection();
-	EntitySystem = Connection->GetSpaceEntitySystem();
-
-	EntitySystem->SetEntityCreatedCallback(
-		[](SpaceEntity* Entity)
-		{
-		});
-
-	bool AssetDetailBlobChangedCallbackCalled = false;
-	csp::common::String CallbackAssetId;
-
-	const csp::common::String ObjectName = "Object 1";
-	SpaceTransform ObjectTransform		 = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-
-	auto [Object] = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-	const csp::common::String ModelAssetId = "NotARealId";
-
-	auto* LightSpaceComponentInstance = (LightSpaceComponent*) Object->AddComponent(ComponentType::Light);
-
-	// Process component creation
-	Object->QueueUpdate();
-	EntitySystem->ProcessPendingEntityOperations();
-
-	// Check component was created
-	auto& Components = *Object->GetComponents();
-	ASSERT_EQ(Components.Size(), 1);
-
-	bool ActionCalled = false;
-	LightSpaceComponentInstance->RegisterActionHandler("TestAction",
-													   [&ActionCalled](ComponentBase*, const csp::common::String, const csp::common::String)
-													   {
-														   ActionCalled = true;
-													   });
-
-	LightSpaceComponentInstance->InvokeAction("TestAction", "TestParam");
-
-	ASSERT_TRUE(ActionCalled);
-
-	// Disconnect from the SignalR server
-	AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_LIGHT_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, LightComponentTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-	auto* AssetSystem	 = SystemsManager.GetAssetSystem();
-
-	const char* TestSpaceName			= "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription	= "OLY-UNITTEST-SPACEDESC-REWIND";
-	const char* TestAssetCollectionName = "OLY-UNITTEST-ASSETCOLLECTION-REWIND";
-	const char* TestAssetName			= "OLY-UNITTEST-ASSET-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	csp::common::String UserId;
-
-	// Log in
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	auto* Connection   = new csp::multiplayer::MultiplayerConnection(Space.Id);
-	auto* EntitySystem = Connection->GetSpaceEntitySystem();
-
-	EntitySystem->SetEntityCreatedCallback(
-		[](SpaceEntity* Entity)
-		{
-		});
-
-	// Connect to the SignalR server
-	auto [Ok] = AWAIT(Connection, Connect);
-
-	EXPECT_TRUE(Ok);
-
-	std::tie(Ok) = AWAIT(Connection, InitialiseConnection);
-
-	EXPECT_TRUE(Ok);
-
-	bool AssetDetailBlobChangedCallbackCalled = false;
-	csp::common::String CallbackAssetId;
-
-	const csp::common::String ObjectName = "Object 1";
-	SpaceTransform ObjectTransform		 = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-
-	auto [Object] = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-	const csp::common::String ModelAssetId = "NotARealId";
-
-	auto* LightSpaceComponentInstance = (LightSpaceComponent*) Object->AddComponent(ComponentType::Light);
-
-	// Process component creation
-	Object->QueueUpdate();
-	EntitySystem->ProcessPendingEntityOperations();
-
-	// Check component was created
-	auto& Components = *Object->GetComponents();
-	ASSERT_EQ(Components.Size(), 1);
-
-	char UniqueAssetCollectionName[256];
-	SPRINTF(UniqueAssetCollectionName, "%s-%s", TestAssetCollectionName, GetUniqueHexString().c_str());
-
-	char UniqueAssetName[256];
-	SPRINTF(UniqueAssetName, "%s-%s", TestAssetName, GetUniqueHexString().c_str());
-
-	// Create asset collection
-	csp::systems::AssetCollection AssetCollection;
-	CreateAssetCollection(AssetSystem, Space.Id, nullptr, UniqueAssetCollectionName, nullptr, nullptr, AssetCollection);
-
-	// Create asset
-	csp::systems::Asset Asset;
-	CreateAsset(AssetSystem, AssetCollection, UniqueAssetName, nullptr, nullptr, Asset);
-	Asset.FileName = "OKO.png";
-	Asset.Name	   = "OKO";
-	Asset.Type	   = csp::systems::EAssetType::IMAGE;
-
-	auto UploadFilePath		 = std::filesystem::absolute("assets/OKO.png");
-	FILE* UploadFile		 = fopen(UploadFilePath.string().c_str(), "rb");
-	uintmax_t UploadFileSize = std::filesystem::file_size(UploadFilePath);
-	auto* UploadFileData	 = new unsigned char[UploadFileSize];
-	fread(UploadFileData, UploadFileSize, 1, UploadFile);
-	fclose(UploadFile);
-
-	csp::systems::BufferAssetDataSource BufferSource;
-	BufferSource.Buffer		  = UploadFileData;
-	BufferSource.BufferLength = UploadFileSize;
-
-	BufferSource.SetMimeType("image/png");
-
-	printf("Uploading asset data...\n");
-
-	// Upload data
-	UploadAssetData(AssetSystem, AssetCollection, Asset, BufferSource, Asset.Uri);
-
-	delete[] UploadFileData;
-
-	EXPECT_EQ(LightSpaceComponentInstance->GetLightCookieType(), LightCookieType::NoCookie);
-	EXPECT_EQ(LightSpaceComponentInstance->GetLightType(), LightType::Point);
-	EXPECT_EQ(LightSpaceComponentInstance->GetLightShadowType(), LightShadowType::None);
-	EXPECT_EQ(LightSpaceComponentInstance->GetInnerConeAngle(), 0.0f);
-	EXPECT_EQ(LightSpaceComponentInstance->GetOuterConeAngle(), 0.78539816339f);
-	EXPECT_EQ(LightSpaceComponentInstance->GetRange(), 1000.0f);
-	EXPECT_EQ(LightSpaceComponentInstance->GetIntensity(), 5000.0f);
-
-	// test values
-	const float InnerConeAngle = 10.0f;
-	const float OuterConeAngle = 20.0f;
-	const float Range		   = 120.0f;
-	const float Intensity	   = 1000.0f;
-
-	LightSpaceComponentInstance->SetLightCookieAssetCollectionId(Asset.AssetCollectionId);
-	LightSpaceComponentInstance->SetLightCookieAssetId(Asset.Id);
-	LightSpaceComponentInstance->SetLightCookieType(LightCookieType::ImageCookie);
-	LightSpaceComponentInstance->SetLightType(LightType::Spot);
-	LightSpaceComponentInstance->SetLightShadowType(LightShadowType::Static);
-	LightSpaceComponentInstance->SetInnerConeAngle(InnerConeAngle);
-	LightSpaceComponentInstance->SetOuterConeAngle(OuterConeAngle);
-	LightSpaceComponentInstance->SetRange(Range);
-	LightSpaceComponentInstance->SetIntensity(Intensity);
-
-	auto LightSpaceComponentKey				= LightSpaceComponentInstance->GetId();
-	auto* StoredLightSpaceComponentInstance = (LightSpaceComponent*) Object->GetComponent(LightSpaceComponentKey);
-
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetLightCookieType(), LightCookieType::ImageCookie);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetLightCookieAssetCollectionId(), Asset.AssetCollectionId);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetLightCookieAssetId(), Asset.Id);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetLightType(), LightType::Spot);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetLightShadowType(), LightShadowType::Static);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetInnerConeAngle(), InnerConeAngle);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetOuterConeAngle(), OuterConeAngle);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetRange(), Range);
-	EXPECT_EQ(StoredLightSpaceComponentInstance->GetIntensity(), Intensity);
-
-	AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_IMAGE_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, ImageComponentTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-	auto* AssetSystem	 = SystemsManager.GetAssetSystem();
-
-	const char* TestSpaceName			= "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription	= "OLY-UNITTEST-SPACEDESC-REWIND";
-	const char* TestAssetCollectionName = "OLY-UNITTEST-ASSETCOLLECTION-REWIND";
-	const char* TestAssetName			= "OLY-UNITTEST-ASSET-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	csp::common::String UserId;
-
-	// Log in
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	auto* Connection   = new csp::multiplayer::MultiplayerConnection(Space.Id);
-	auto* EntitySystem = Connection->GetSpaceEntitySystem();
-
-	EntitySystem->SetEntityCreatedCallback(
-		[](SpaceEntity* Entity)
-		{
-		});
-
-	// Connect to the SignalR server
-	auto [Ok] = AWAIT(Connection, Connect);
-
-	EXPECT_TRUE(Ok);
-
-	std::tie(Ok) = AWAIT(Connection, InitialiseConnection);
-
-	EXPECT_TRUE(Ok);
-
-	bool AssetDetailBlobChangedCallbackCalled = false;
-	csp::common::String CallbackAssetId;
-
-	const csp::common::String ObjectName = "Object 1";
-	SpaceTransform ObjectTransform		 = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-
-	auto [Object] = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-	const csp::common::String ModelAssetId = "NotARealId";
-
-	auto* ImageSpaceComponentInstance = (ImageSpaceComponent*) Object->AddComponent(ComponentType::Image);
-
-	// Process component creation
-	Object->QueueUpdate();
-	EntitySystem->ProcessPendingEntityOperations();
-
-	// Check component was created
-	auto& Components = *Object->GetComponents();
-	ASSERT_EQ(Components.Size(), 1);
-
-	char UniqueAssetCollectionName[256];
-	SPRINTF(UniqueAssetCollectionName, "%s-%s", TestAssetCollectionName, GetUniqueHexString().c_str());
-
-	char UniqueAssetName[256];
-	SPRINTF(UniqueAssetName, "%s-%s", TestAssetName, GetUniqueHexString().c_str());
-
-	// Create asset collection
-	csp::systems::AssetCollection AssetCollection;
-	CreateAssetCollection(AssetSystem, Space.Id, nullptr, UniqueAssetCollectionName, nullptr, nullptr, AssetCollection);
-
-	// Create asset
-	csp::systems::Asset Asset;
-	CreateAsset(AssetSystem, AssetCollection, UniqueAssetName, nullptr, nullptr, Asset);
-	Asset.FileName = "OKO.png";
-	Asset.Name	   = "OKO";
-	Asset.Type	   = csp::systems::EAssetType::IMAGE;
-
-	auto UploadFilePath		 = std::filesystem::absolute("assets/OKO.png");
-	FILE* UploadFile		 = fopen(UploadFilePath.string().c_str(), "rb");
-	uintmax_t UploadFileSize = std::filesystem::file_size(UploadFilePath);
-	auto* UploadFileData	 = new unsigned char[UploadFileSize];
-	fread(UploadFileData, UploadFileSize, 1, UploadFile);
-	fclose(UploadFile);
-
-	csp::systems::BufferAssetDataSource BufferSource;
-	BufferSource.Buffer		  = UploadFileData;
-	BufferSource.BufferLength = UploadFileSize;
-
-	BufferSource.SetMimeType("image/png");
-
-	printf("Uploading asset data...\n");
-
-	// Upload data
-	UploadAssetData(AssetSystem, AssetCollection, Asset, BufferSource, Asset.Uri);
-
-	delete[] UploadFileData;
-
-	EXPECT_EQ(ImageSpaceComponentInstance->GetBillboardMode(), BillboardMode::Off);
-	EXPECT_EQ(ImageSpaceComponentInstance->GetDisplayMode(), DisplayMode::DoubleSided);
-	EXPECT_EQ(ImageSpaceComponentInstance->GetIsARVisible(), true);
-	EXPECT_EQ(ImageSpaceComponentInstance->GetIsEmissive(), false);
-
-
-	ImageSpaceComponentInstance->SetAssetCollectionId(Asset.AssetCollectionId);
-	ImageSpaceComponentInstance->SetImageAssetId(Asset.Id);
-	ImageSpaceComponentInstance->SetBillboardMode(BillboardMode::YawLockedBillboard);
-	ImageSpaceComponentInstance->SetDisplayMode(DisplayMode::SingleSided);
-	ImageSpaceComponentInstance->SetIsARVisible(false);
-	ImageSpaceComponentInstance->SetIsEmissive(true);
-
-	auto ImageSpaceComponentKey		= ImageSpaceComponentInstance->GetId();
-	auto* StoredImageSpaceComponent = (ImageSpaceComponent*) Object->GetComponent(ImageSpaceComponentKey);
-
-	EXPECT_EQ(StoredImageSpaceComponent->GetAssetCollectionId(), Asset.AssetCollectionId);
-	EXPECT_EQ(StoredImageSpaceComponent->GetImageAssetId(), Asset.Id);
-	EXPECT_EQ(StoredImageSpaceComponent->GetBillboardMode(), BillboardMode::YawLockedBillboard);
-	EXPECT_EQ(StoredImageSpaceComponent->GetDisplayMode(), DisplayMode::SingleSided);
-	EXPECT_EQ(StoredImageSpaceComponent->GetIsARVisible(), false);
-	EXPECT_EQ(StoredImageSpaceComponent->GetIsEmissive(), true);
-
-	AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_REFLECTION_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, ReflectionComponentTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-	auto* AssetSystem	 = SystemsManager.GetAssetSystem();
-
-	const char* TestSpaceName			= "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription	= "OLY-UNITTEST-SPACEDESC-REWIND";
-	const char* TestAssetCollectionName = "OLY-UNITTEST-ASSETCOLLECTION-REWIND";
-	const char* TestAssetName			= "OLY-UNITTEST-ASSET-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	csp::common::String UserId;
-
-	// Log in
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	auto* Connection   = new csp::multiplayer::MultiplayerConnection(Space.Id);
-	auto* EntitySystem = Connection->GetSpaceEntitySystem();
-
-	EntitySystem->SetEntityCreatedCallback(
-		[](SpaceEntity* Entity)
-		{
-		});
-
-	// Connect to the SignalR server
-	auto [Ok] = AWAIT(Connection, Connect);
-
-	EXPECT_TRUE(Ok);
-
-	std::tie(Ok) = AWAIT(Connection, InitialiseConnection);
-
-	EXPECT_TRUE(Ok);
-
-	bool AssetDetailBlobChangedCallbackCalled = false;
-	csp::common::String CallbackAssetId;
-
-	const csp::common::String ObjectName = "Object 1";
-	SpaceTransform ObjectTransform		 = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-
-	auto [Object] = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-	const csp::common::String ModelAssetId = "NotARealId";
-
-	auto* ReflectionSpaceComponentInstance = (ReflectionSpaceComponent*) Object->AddComponent(ComponentType::Reflection);
-
-	// Process component creation
-	Object->QueueUpdate();
-	EntitySystem->ProcessPendingEntityOperations();
-
-	// Check component was created
-	auto& Components = *Object->GetComponents();
-	ASSERT_EQ(Components.Size(), 1);
-
-	char UniqueAssetCollectionName[256];
-	SPRINTF(UniqueAssetCollectionName, "%s-%s", TestAssetCollectionName, GetUniqueHexString().c_str());
-
-	char UniqueAssetName[256];
-	SPRINTF(UniqueAssetName, "%s-%s", TestAssetName, GetUniqueHexString().c_str());
-
-	// Create asset collection
-	csp::systems::AssetCollection AssetCollection;
-	CreateAssetCollection(AssetSystem, Space.Id, nullptr, UniqueAssetCollectionName, nullptr, nullptr, AssetCollection);
-
-	// Create asset
-	csp::systems::Asset Asset;
-	CreateAsset(AssetSystem, AssetCollection, UniqueAssetName, nullptr, nullptr, Asset);
-	Asset.FileName = "OKO.png";
-	Asset.Name	   = "OKO";
-	Asset.Type	   = csp::systems::EAssetType::IMAGE;
-
-	auto UploadFilePath		 = std::filesystem::absolute("assets/OKO.png");
-	FILE* UploadFile		 = fopen(UploadFilePath.string().c_str(), "rb");
-	uintmax_t UploadFileSize = std::filesystem::file_size(UploadFilePath);
-	auto* UploadFileData	 = new unsigned char[UploadFileSize];
-	fread(UploadFileData, UploadFileSize, 1, UploadFile);
-	fclose(UploadFile);
-
-	csp::systems::BufferAssetDataSource BufferSource;
-	BufferSource.Buffer		  = UploadFileData;
-	BufferSource.BufferLength = UploadFileSize;
-
-	BufferSource.SetMimeType("image/png");
-
-	printf("Uploading asset data...\n");
-
-	// Upload data
-	UploadAssetData(AssetSystem, AssetCollection, Asset, BufferSource, Asset.Uri);
-
-	delete[] UploadFileData;
-
-	EXPECT_EQ(ReflectionSpaceComponentInstance->GetReflectionShape(), ReflectionShape::UnitBox);
-
-	ReflectionSpaceComponentInstance->SetAssetCollectionId(Asset.AssetCollectionId);
-	ReflectionSpaceComponentInstance->SetReflectionAssetId(Asset.Id);
-	ReflectionSpaceComponentInstance->SetReflectionShape(ReflectionShape::UnitSphere);
-
-	auto ReflectionSpaceComponentKey	 = ReflectionSpaceComponentInstance->GetId();
-	auto* StoredReflectionSpaceComponent = (ReflectionSpaceComponent*) Object->GetComponent(ReflectionSpaceComponentKey);
-
-	EXPECT_EQ(StoredReflectionSpaceComponent->GetAssetCollectionId(), Asset.AssetCollectionId);
-	EXPECT_EQ(StoredReflectionSpaceComponent->GetReflectionAssetId(), Asset.Id);
-	EXPECT_EQ(StoredReflectionSpaceComponent->GetReflectionShape(), ReflectionShape::UnitSphere);
-
-	AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
-
 // Derived type that allows us to access protected members of SpaceEntitySystem
 struct InternalSpaceEntitySystem : public csp::multiplayer::SpaceEntitySystem
 {
@@ -3711,96 +1783,6 @@ struct InternalSpaceEntitySystem : public csp::multiplayer::SpaceEntitySystem
 		Entities.Clear();
 	}
 };
-
-#if RUN_ALL_UNIT_TESTS || RUN_MULTIPLAYER_TESTS || RUN_MULTIPLAYER_FOG_COMPONENT_TEST
-CSP_PUBLIC_TEST(CSPEngine, MultiplayerTests, FogComponentTest)
-{
-	SetRandSeed();
-
-	auto& SystemsManager = csp::systems::SystemsManager::Get();
-	auto* UserSystem	 = SystemsManager.GetUserSystem();
-	auto* SpaceSystem	 = SystemsManager.GetSpaceSystem();
-
-	const char* TestSpaceName		 = "OLY-UNITTEST-SPACE-REWIND";
-	const char* TestSpaceDescription = "OLY-UNITTEST-SPACEDESC-REWIND";
-
-	char UniqueSpaceName[256];
-	SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueHexString().c_str());
-
-	// Log in
-	csp::common::String UserId;
-	LogIn(UserSystem, UserId);
-
-	// Create space
-	csp::systems::Space Space;
-	CreateSpace(SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, Space);
-
-	auto [EnterResult] = AWAIT_PRE(SpaceSystem, EnterSpace, RequestPredicate, Space.Id, true);
-	EXPECT_EQ(EnterResult.GetResultCode(), csp::services::EResultCode::Success);
-	Connection	 = EnterResult.GetConnection();
-	EntitySystem = Connection->GetSpaceEntitySystem();
-
-	EntitySystem->SetEntityCreatedCallback(
-		[](SpaceEntity* Entity)
-		{
-		});
-
-	// Create object to represent the fog
-	csp::common::String ObjectName = "Object 1";
-	SpaceTransform ObjectTransform = {csp::common::Vector3::Zero(), csp::common::Vector4::Zero(), csp::common::Vector3::One()};
-	auto [CreatedObject]		   = AWAIT(EntitySystem, CreateObject, ObjectName, ObjectTransform);
-
-	// Create fog component
-	auto* FogComponent = static_cast<FogSpaceComponent*>(CreatedObject->AddComponent(ComponentType::Fog));
-
-	// Ensure defaults are set
-	EXPECT_EQ(FogComponent->GetFogMode(), FogMode::Linear);
-	EXPECT_EQ(FogComponent->GetPosition(), csp::common::Vector3::Zero());
-	EXPECT_EQ(FogComponent->GetRotation(), csp::common::Vector4(0, 0, 0, 1));
-	EXPECT_EQ(FogComponent->GetScale(), csp::common::Vector3::One());
-	EXPECT_FLOAT_EQ(FogComponent->GetStartDistance(), 0.f);
-	EXPECT_FLOAT_EQ(FogComponent->GetEndDistance(), 0.f);
-	EXPECT_EQ(FogComponent->GetColor(), csp::common::Vector3({0.8f, 0.9f, 1.0f}));
-	EXPECT_FLOAT_EQ(FogComponent->GetDensity(), 0.2f);
-	EXPECT_FLOAT_EQ(FogComponent->GetHeightFalloff(), 0.2f);
-	EXPECT_FLOAT_EQ(FogComponent->GetMaxOpacity(), 1.f);
-	EXPECT_FALSE(FogComponent->GetIsVolumetric());
-
-	// Set new values
-	FogComponent->SetFogMode(FogMode::Exponential);
-	FogComponent->SetPosition(csp::common::Vector3::One());
-	FogComponent->SetRotation(csp::common::Vector4(0, 0, 0, 1));
-	FogComponent->SetScale(csp::common::Vector3(2, 2, 2));
-	FogComponent->SetStartDistance(1.1f);
-	FogComponent->SetEndDistance(2.2f);
-	FogComponent->SetColor(csp::common::Vector3::One());
-	FogComponent->SetDensity(3.3f);
-	FogComponent->SetHeightFalloff(4.4f);
-	FogComponent->SetMaxOpacity(5.5f);
-	FogComponent->SetIsVolumetric(true);
-
-	// Ensure values are set correctly
-	EXPECT_EQ(FogComponent->GetFogMode(), FogMode::Exponential);
-	EXPECT_EQ(FogComponent->GetPosition(), csp::common::Vector3::One());
-	EXPECT_EQ(FogComponent->GetRotation(), csp::common::Vector4(0, 0, 0, 1));
-	EXPECT_EQ(FogComponent->GetScale(), csp::common::Vector3(2, 2, 2));
-	EXPECT_FLOAT_EQ(FogComponent->GetStartDistance(), 1.1f);
-	EXPECT_FLOAT_EQ(FogComponent->GetEndDistance(), 2.2f);
-	EXPECT_EQ(FogComponent->GetColor(), csp::common::Vector3::One());
-	EXPECT_FLOAT_EQ(FogComponent->GetDensity(), 3.3f);
-	EXPECT_FLOAT_EQ(FogComponent->GetHeightFalloff(), 4.4f);
-	EXPECT_FLOAT_EQ(FogComponent->GetMaxOpacity(), 5.5f);
-	EXPECT_TRUE(FogComponent->GetIsVolumetric());
-
-	AWAIT(SpaceSystem, ExitSpaceAndDisconnect, Connection);
-
-	// Delete space
-	DeleteSpace(SpaceSystem, Space.Id);
-
-	// Log out
-	LogOut(UserSystem);
-}
-#endif
 
 // Disabled by default as it can be slow
 #if RUN_MULTIPLAYER_MANYENTITIES_TEST
